enable flagged overwriting of share_profile, fixed ev.sh

+++ bin/setup.sh
@@ -1,5 +1,21 @@
 #!/bin/sh
 
+print_usage() {
+  printf "Usage: setup.sh [c|n]"
+}
+
+# only for 
+# while getopts 'abc:x:yz' flag; do
+while getopts 'n' flag; do
+  case "${flag}" in
+    n) new_profile='true' ;;
+    ## c) c_arg="${OPTARG}" ;;
+    ## x) x_arg="${OPTARG}" ;;
+    *) print_usage
+       # exit 1 ;;
+  esac
+done
+
 . ./include/env.sh
 
 cd "$CWH/bin"
@@ -13,9 +29,9 @@ if [ ! -d "$CLEAN_DIR" ]; then
 	mkdir "$CLEAN_DIR"
 fi
 
-if [ ! -d "$SHARE_DIR" ]; then
-	mkdir "$SHARE_DIR"
-fi
+# if [ ! -d "$SHARE_DIR" ]; then
+# 	mkdir "$SHARE_DIR"
+# fi
 
 # redundant, but for clarity safe
 if [ ! -d "$SHARE_DIR/bin" ]; then
@@ -34,18 +50,25 @@ fi
 # remove existing if any
 share_file="$SHARE_DIR/etc/.share_profile"
 
-if [ -f $share_file ]; then
+if [ -f $share_file ] && [ -n "$new_profile" ]; then
 	echo "Removing user share_profile"
-
 	rm "$SHARE_DIR/etc/.share_profile"
+else
+    echo "Reuse existing share_profile"
 fi
 
-echo "Building new share_profile"
-printf "%s\n%s\n" "SHARE_DIR=\"$SHARE_DIR\"" "$(cat $CWH/etc/.share_profile)" > "$SHARE_DIR/etc/.share_profile"
+echo "Using new share_profile? ..."
+
+if [ -e "$share_file" ]; then
+    echo "...Nope"
+else
+    echo "..Yup"
+    cp "$CWH/etc/.share_profile"  "$SHARE_DIR/etc/.share_profile"
+fi
 
 # include share_profile in .profile
 #match="$(cat $HOME/.profile | grep '.share_profile')"
-match="$(cat $HOME/.profile | egrep -o "\s?\.\s[\"\']?$SHARE_DIR/etc/.share_profile[\"\']")"
+match="$([ -f $HOME/.profile ] && cat $HOME/.profile | egrep -o "\s?\.\s[\"\']?$SHARE_DIR/etc/.share_profile[\"\']")"
 
 if [ -z "$match" ] && [ -f "$HOME/.profile" ]; then 
 	cp "$HOME/.profile" "$CLEAN_DIR/.profile"
@@ -57,7 +80,7 @@ fi
 
 # include share_profile in .bash_profile
 # match2="$(cat $HOME/.bash_profile | grep '.share_profile')"
-match2="$(cat $HOME/.bash_profile | egrep -o "\s?\.\s[\"\']?$SHARE_DIR/etc/.share_profile[\"\']")"
+match2="$([ -f $HOME/.bash_profile ] && cat $HOME/.bash_profile | egrep -o "\s?\.\s[\"\']?$SHARE_DIR/etc/.share_profile[\"\']")"
                                   
 if [ -z "$match2" ] && [ -f "$HOME/.bash_profile" ]; then 
 	cp "$HOME/.bash_profile" "$CLEAN_DIR/.bash_profile"
@@ -69,7 +92,7 @@ fi
 
 # include share_profile in .bashrc
 # match3="$(cat $HOME/.bashrc | grep '.share_profile')"
-match3="$(cat $HOME/.bashrc | egrep -o "\s?\.\s[\"\']?$SHARE_DIR/etc/.share_profile[\"\']")"
+match3="$([ -f $HOME/.bashrc ] && cat $HOME/.bashrc | egrep -o "\s?\.\s[\"\']?$SHARE_DIR/etc/.share_profile[\"\']")"
 
 if [ -z "$match3" ] && [ -f "$HOME/.bashrc" ]; then
 	cp "$HOME/.bashrc" "$CLEAN_DIR/.bashrc"
@@ -79,6 +102,7 @@ if [ -z "$match3" ] && [ -f "$HOME/.bashrc" ]; then
 	printf "\n%s\n%s\n" "# include .share_profile if it exists" "[[ -f \"$SHARE_DIR/etc/.share_profile\" ]] && . \"$SHARE_DIR/etc/.share_profile\"" >> $HOME/.bashrc
 fi
 
+# bash completion
 default_inputrc="/etc/inputrc"
 
 if [ -f "$default_inputrc" ]; then
